backend/requirements.txt
asgiref==3.3.4
celery==5.2.3
Django==3.2.5
django-cors-headers==3.10.1
django-cprofile-middleware==1.0.5
django-redis==5.2.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
flower==1.0.0
gunicorn==20.1.0
imageio==2.13.5
imutils==0.5.4
keras==2.7.0
kombu==5.2.3
locust==2.5.1
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.5.62
pandas==1.3.5
Pillow==9.0.0
psutil==5.9.0
psycopg2-binary==2.9.3
redis==4.1.0
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.5.1
tensorflow-cpu==2.7.0
tensorboard==2.7.0
uvicorn==0.15.0
daphne==3.0.2